pyproject.toml
[project]
name = "crystal-ballin-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.18,<3.10"
dependencies = [
    "flask>=3.1.0",
    "matplotlib>=3.9.3",
    "pandas>=2.2.3",
    "prophet>=1.1.6",
    "python-dotenv>=1.0.1",
    "yfinance>=0.2.50",
]
